+++ Code/Comscore-Examples/Cocoapod/ComscoreReporter/ComscoreConfig.swift
@@ -40,6 +40,8 @@ let comscoreConfig = ComScoreConfiguration(
     publisherId: "put your publisher id here",
     applicationName: "put your application name here",
     userConsent: .granted,
+    usagePropertiesAutoUpdateMode: .foregroundOnly,
+    usagePropertiesAutoUpdateInterval: 90,
     adIdProcessor: nil,
     debug: true
 )