+++ Code/Comscore/Source/integration/ComscoreStreamingAnalytics.swift
@@ -6,6 +6,7 @@
 
 import Foundation
 import THEOplayerSDK
+import ComScore
 
 public class ComScoreStreamingAnalytics {
     let adapter: THEOComScoreAdapter
@@ -62,6 +63,35 @@ public class ComScoreStreamingAnalytics {
      */
     public func setPersistentLabels(labels: [String: String]) {
         adapter.setPersistentLabels(labels: labels)
+    }
+    
+    /**
+     Report that your application starts providing the user experience
+     */
+    func notifyUxActive() {
+        SCORAnalytics.notifyUxActive()
+    }
+
+    /**
+     Report that your application stops providing the user experience. E.g. when the application moves to the background and the user turned off
+     a background audio or picture-in-picture capability
+     */
+    func notifyUxInactive() {
+        SCORAnalytics.notifyUxInactive()
+    }
+
+    /**
+     Report that your application moved to the foreground
+     */
+    func notifyEnterForeground() {
+        SCORAnalytics.notifyEnterForeground()
+    }
+
+    /**
+     Report that your application moved to the background
+     */
+    func notifyExitForeground() {
+        SCORAnalytics.notifyExitForeground()
     }
     
 //    /**