+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Comscore
+  - Added `usagePropertiesAutoUpdateMode` and `usagePropertiesAutoUpdateInterval` to `ComScoreConfiguration`.
+  - Exposed `notifyUxActive`, `notifyUxInactive`, `notifyEnterForeground` and `notifyExitForeground` methods.
+
 ## [8.7.0] - 2024-12-18
 
 ### Added